Posted Website
This is the posted website. Validates with no errors. Responsive desktop and mobile views. There is still a lot of work to be done; especially in mobile view. The images need to be formatted for mobile view (smaller file size). Need to finalize the form. Sermons page needs more styling.